refactor(client-ui): refactor chat sidebar for modularity
