+++ apps/client-ui/src/components/chat/chat-history-date-group/chat-history-date-group.tsx
@@ -1,0 +1,20 @@
+interface ChatHistoryDateGroupProps {
+	dateGroup: string
+	children: React.ReactNode
+}
+
+const ChatHistoryDateGroup = ({
+	dateGroup,
+	children,
+}: ChatHistoryDateGroupProps) => {
+	return (
+		<div key={dateGroup} className="py-2">
+			<div className="px-3 py-2 text-xs text-muted-foreground font-medium">
+				{dateGroup}
+			</div>
+			<div className="space-y-1">{children}</div>
+		</div>
+	)
+}
+
+export { ChatHistoryDateGroup }
